backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
opencv-python==4.9.0.80
numpy>=1.26.0
Pillow==10.2.0
pytesseract==0.3.10
requests>=2.31.0
imageio-ffmpeg>=0.5.0
mediapipe>=0.10.0
mediapipe>=0.10.0
imageio-ffmpeg>=0.5.0
ultralytics>=8.3.0
lap>=0.5.12
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=7.0.0
